stricter type

+++ src/SPECK_Storage.h
@@ -57,7 +57,7 @@ public:
     // and what the volume/slice dimension is in num. of elements.
     // Note: don't need to provide the buffer size because this function
     //       goes to a fixed location to retrieve the stream size.
-    auto get_speck_stream_size( const void* ) const -> size_t;
+    auto get_speck_stream_size( const void* ) const -> uint64_t;
     auto get_speck_stream_dims( const void* ) const -> std::array<size_t, 3>;
 
 protected:
